refactor: 임시로 사용한 TEMP_를 제거한 이름으로 리네임
