+++ components/templates/LibrarySelectionMapTemplate.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { MapBound } from "@/types/MapBound";
 import LibraryMarkerInfo from "@/types/LibraryMarkerInfo";
-import { LibrarySelectionMap } from "../organisms/map/TEMP_LibraryMap";
+import { LibrarySelectionMap } from "../organisms/map/LibraryMap";
 
 export interface Props {
   mapBound: MapBound;
